chore(TSC): :fire: Cleaned up the proj's tsconig.json file by swapping some props, formatting the doc, & removing unnesccary mod resolution support
Included in the tsconfig clean up was the removal of the "files" property, as the files list was getting long. I added the much more tidy includes property. Also I added an excludes property as a safe guard if any proj structure or include/exclude settings change in the fututre. Removed the config for supporting "Json as a module resolution", the project doesn't use it, and will likely never use it. Formatted the document as well.