test: :white_check_mark: Added test generator for HSV class, changed hsv classes file name, and hsv test file's name.
Changed the name of of api/color.ts to api/hsv.ts  because the project's color class is going to be broken up into multiple classes. The file now contains the HSV class. Finished the HSV class, and changed its file name to hsv.test.ts to match the hsv class file name.