test(TESTING): :hammer: Added assertion for validating HSV to RGB conversions
HSV doesn't always convert to the same RGB value that was used to create it, and HSV algorithms vary, all this is attributed to the rounding that is neccassary to impliment in the algorithm, therefore a special assertion is needed to find if the RGB emitted by the conversion is with in (Red 1+/-),   (Green 1+/-)  &  (Blue 1+/-) values of the expected rgb value.

I don't expect this to break any HSV emitted rgb comparisons, but if compared against HSV from a 3rd party source the compares may fail. In such a case more info about the 3rd part algorithm would be needed to solve the issue.